[ADD, MOD] Created an AnimUtils class to manage animation functionality. Now includes a coroutine method that halts a script until animation is complete.
